src/requirements.txt
asgiref==3.4.1
Django==3.2.5
dj-database-url==0.5.0
django-extensions==3.1.1
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.0.4
whitenoise==5.2.0
snips-nlu==0.20.2